requirements.txt
fastapi
psycopg2-binary
python-dotenv
uvicorn
langchain
langchain-community
langchain-ollama
langchain-groq
sqlalchemy
